Add clang-tidy checks

+++ src/bindings/key.cpp
@@ -2,6 +2,8 @@
 
 #include <sol/sol.hpp>
 
+#include "../typedef/key_mod.hpp"
+
 void Key::init_bridge(sol::table& core) {
     // clang-format off
     core.new_usertype<Key>("Key",
@@ -10,7 +12,10 @@ void Key::init_bridge(sol::table& core) {
         "to_string", &Key::to_string,
         /// Normalizes a Key string.
         "normalize", [](const std::string_view str) -> std::string {
-            if (Key key{0, KeyMod::NONE}; try_parse_string(str, key)) { return key.to_string(); }
+            if (Key key{0, static_cast<std::size_t>(KeyMod::NONE)}; try_parse_string(str, key)) {
+                return key.to_string();
+            }
+
             return std::string(str);
         });
     // clang-format on
